Moved check_signal syntax to cpp file
